Creating Engine
The engine will get input from UI (not implemented yet) and store FileChecker objects in lists

The delete_file method will check and delete the file in the list if the timing is right.

One thing to not is that deleting files wont delete the FileChecker object in the list, because I might use them to restore maybe or get deletion history.